Fase 3/APIs/DB subscriber/requirements.txt
paho-mqtt
firebase-admin
pytz